OS/2.
git-svn-id: https://svn.code.sf.net/p/spot-on/code/branches/0.05@1880 0a995953-517f-4c62-b348-104708d926b2

+++ Kernel/spot-on-listener.cc
@@ -55,7 +55,7 @@ void spoton_listener_tcp_server::incomingConnection(int socketDescriptor)
 	  QHostAddress address;
 	  sockaddr nativeAddress;
 #ifdef Q_OS_OS2
-	  size_t length = sizeof(nativeAddress);
+	  int length = sizeof(nativeAddress);
 #else
 	  socklen_t length = sizeof(nativeAddress);
 #endif
